Fix various compilation warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@132479 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: dedd7cad9d83dabf17c95a937d7bd52392d89882

+++ Rec/Rich/RichGlobalPID/src/RichGlobalPIDTrTrackSel.cpp
@@ -115,7 +115,7 @@ StatusCode TrackSel::eventInit()
     deleteGPIDEvent();
     return Warning("No tracks selected for processing -> Stopped",StatusCode::SUCCESS,0);
   }
-  else if ( (int)richTracks()->size() > m_maxUsedTracks )
+  else if ( richTracks()->size() > m_maxUsedTracks )
   {
     procStatus()->addAlgorithmStatus( gpidName(), "RICH", "ReachedRichTrackLimit",
                                       Rich::Rec::ReachedRichTrackLimit, true );
